+++ src/functions.php
@@ -8,6 +8,7 @@ require get_template_directory() . '/inc/customizer.php';
 require get_template_directory() . '/inc/template-tags.php';
 require get_template_directory() . '/inc/github-updater.php';
 require get_template_directory() . '/inc/qrcode-generator.php';
+require get_template_directory() . '/inc/shortcodes.php';
 
 // Theme-Setup
 function potsdam_rechtsanwalt_setup() {